refactor: Renamed django app 'todo_app' to 'tasks' according to mentor suggestion
