+++ tasks/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from todo_app.models import Tag, Task
+from tasks.models import Tag, Task
 
 admin.site.register(Tag)
 admin.site.register(Task)
