+++ tasks/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class TodoAppConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "todo_app"
+    name = "tasks"
