+++ templates/includes/sidebar.html
@@ -1,4 +1,4 @@
 <ul class="sidebar-nav list-group">
-  <li class="list-group-item"><a href="{% url 'todo_app:index' %}">Home</a></li>
-  <li class="list-group-item"><a href="{% url 'todo_app:tag-list' %}">Tags</a></li>
+  <li class="list-group-item"><a href="{% url 'tasks:index' %}">Home</a></li>
+  <li class="list-group-item"><a href="{% url 'tasks:tag-list' %}">Tags</a></li>
 </ul>
