+++ todo_list/settings.py
@@ -37,7 +37,7 @@ INSTALLED_APPS = [
     "django.contrib.staticfiles",
     "crispy_forms",
     "crispy_bootstrap4",
-    "todo_app"
+    "tasks"
 ]
 
 MIDDLEWARE = [
